Ui adjustments, moved header wrapper to helper folder
