+++ public/components/integrations/components/integration_header.tsx
@@ -22,7 +22,7 @@ import {
   OPENSEARCH_DOCUMENTATION_URL,
 } from '../../../../common/constants/integrations';
 import { IntegrationUploadFlyout } from './upload_flyout';
-import { HeaderControlledComponentsWrapper } from '../../../../public/plugin_headerControl';
+import { HeaderControlledComponentsWrapper } from '../../../../public/plugin_helpers/plugin_headerControl';
 import { coreRefs } from '../../../framework/core_refs';
 
 const newNavigation = coreRefs.chrome?.navGroup.getNavGroupEnabled();
