+++ public/components/overview/index.scss
@@ -1,0 +1,10 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+.obsOverviewDashboardHeader
+{
+    line-height: 32px;
+    white-space: nowrap;
+}
